Add force_destroy option & default lifecycle rule
In order for Terratest to work we need to be able to destroy the bucket.

To do this we have removed the lifecycle rule preventing destroy and
have added a new variable to set the force destroy option.  This is
needed as some other modules such as the bastion which use this module
create objects in the bucket.  This will also be useful for users who
wish to destroy buckets.

Added a default lifecyle rule for object management, this was listed in
the inputs as not required but was actually.  Using the rule in the
usage example as the sensible default.