Add AES256 encryption for sse_algorithm
S3 Access bucket logging requires AES256 encryption to work. This change
is to add this to the module so that the module can be used when S3
bucket logging is required.

Tests added to confirm that the logic for the sse algorithm works, along
with confirming that the default remains `aws:kms` to ensure there are
no breaking changes.

Removed unneed code and resolved SCA checks.